+++ gradio/app.py
@@ -10,7 +10,7 @@ from typing import Any, Dict, List
 import numpy as np
 import numpy.typing as npt
 from datasets import load_dataset
-from pollen_vision.perception.utils import Annotator, get_bboxes
+from pollen_vision.utils import Annotator, get_bboxes
 from pollen_vision.vision_models.object_detection import OwlVitWrapper
 from pollen_vision.vision_models.object_segmentation import MobileSamWrapper
