+++ scripts/annotate_video.py
@@ -23,7 +23,7 @@ import asyncio
 import cv2
 import numpy as np
 import tqdm
-from pollen_vision.perception.utils import Annotator, get_bboxes
+from pollen_vision.utils import Annotator, get_bboxes
 from pollen_vision.vision_models.object_detection import OwlVitWrapper
 from pollen_vision.vision_models.object_segmentation import MobileSamWrapper
 from recorder import Recorder
